pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = [""]
include = ["pyserini*"]
exclude = ["tests", "integrations"]
namespaces = false

[tool.setuptools.package-data]
"pyserini.server.rest" = ["openapi.yaml"]

[tool.black]
skip-string-normalization = true

[project]
name = "pyserini"
version = "2.0.0"
license = "Apache-2.0"
license-files = ["LICENSE.txt",
]
authors = [
    { name="Jimmy Lin", email="jimmylin@uwaterloo.ca" },
]
description = "A Python toolkit for reproducible information retrieval research with sparse and dense representations"
readme = "project-description.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">=3.12"
dependencies = [
    "datasets",
    "fastapi>=0.124",
    "fastmcp>=2.14",
    "flask",
    "matplotlib",
    "numpy>=2",
    "onnxruntime>=1.23",
    "openai>=2.12.0",
    "pandas",
    "pillow>=12",
    "pyjnius>=1.7.0",
    "pyyaml",
    "requests",
    "scikit-learn",
    "scipy",
    "sentencepiece>=0.2",
    "tiktoken>=0.12",
    "torch>=2.9",
    "tqdm",
    "transformers>=4.57",
    "uvicorn>=0.38"
]

[project.urls]
Homepage = "https://github.com/castorini/pyserini"
Issues = "https://github.com/castorini/pyserini/issues"

[project.optional-dependencies]
optional = [
    "uniir-for-pyserini>=0.1.3",
    "vlm2vec-for-pyserini>=0.2.0"
]
